BackupTableinFileMaker
AppleScript for execution in FileMaker to :
Part 1. Backup whole database to zip on Desktop with filename = databaseName
Part 2. Identify and create a unique name for output file, based on YYYY-MM-DD timestamp and databaseName
Part 3. Copy zip to a cloud service for offsite backup ; copy file to disk2 and disk3 for redundant on-site storage ; if disks not connected, new folder created on desktop to temporarily store files until drives connected. 
On completion, file is backed up to 2 external devices on-site, as well as one cloud service (and potentially a second, depending on location of primary file)