Create MakeFoldersInFinder
Sometimes, you need a number of folders with specific names to be created, to provide specific destinations for specific files; one quick and easy way to do this is with an AppleScript repeat loop. In this case, we create 20 folders where we can store the job applications (for example) for 20 anonymised candidates.